Add AuthZ service support
An additional security control for AuthZ has been added. While disabled
(the default), all Google Drive documents are present in the context.
When enabled, only documents that the user has access to are present.

Propagating Google Drive permissions to AuthZ is done via a script. A
GitHub workflow has also been added to make it possible to update the
tuples without needing to checkout the code.