Send entire LLM input to AI Guard
Previously we were only sending the user's prompt to AI Guard, and
indeed this was fine when we were only interested in demoing the guards
on malicious or sensitive user input. However, the next demos want to
craft attacks that involve the system prompt and the RAG documents, so
AI Guard needs to cover everything that's going to the LLM.

The easiest way to accomplish this would be to run AI Guard in
`POST /api/ai` before sending the input to the LLM, and this is probably
how it would be done in a real world application. However, this chat app
runs the AI Guard check as a separate step with its own API endpoint
because we want to keep it toggleable with its own processing step and
loading indicator for demonstration purposes. So to accommodate this, a
slightly bigger refactor is done:

1. Document fetching is now done in a new endpoint in order to have the
  documents earlier than chat completions generation time.
2. When an input is submitted, the documents are fetched and the system
  prompt, user prompt, and documents are all collected and sent to AI
  Guard.
3. Guarded input is sent to `POST /api/ai`.